+++ docs/source/conf.py
@@ -29,7 +29,7 @@ author = 'BIH Core Unit Bioinformatics'
 # The short X.Y version
 version = '0.10'
 # The full version, including alpha/beta/rc tags
-release = '0.10.8'
+release = '0.10.9-WIP'
 
 
 # -- General configuration ---------------------------------------------------