+++ projectroles/tests/test_views_ajax.py
@@ -1,6 +1,9 @@
 """Ajax API view tests for the projectroles app"""
 
+import json
+
 from django.forms import model_to_dict
+from django.test import override_settings
 from django.urls import reverse
 
 from projectroles.models import ProjectUserTag, PROJECT_TAG_STARRED
@@ -9,13 +12,96 @@ from projectroles.tests.test_models import (
     RoleAssignmentMixin,
     ProjectUserTagMixin,
 )
-from projectroles.tests.test_views import TestViewsBase, PROJECT_TYPE_PROJECT
+from projectroles.tests.test_views import (
+    TestViewsBase,
+    PROJECT_TYPE_CATEGORY,
+    PROJECT_TYPE_PROJECT,
+)
+
+
+class TestProjectListColumnAjaxView(
+    ProjectMixin, RoleAssignmentMixin, ProjectUserTagMixin, TestViewsBase
+):
+    """Tests for ProjectListColumnAjaxView"""
+
+    def setUp(self):
+        super().setUp()
+        self.category = self._make_project(
+            'TestCategory', PROJECT_TYPE_CATEGORY, None
+        )
+        self.owner_as_cat = self._make_assignment(
+            self.category, self.user, self.role_owner
+        )
+        self.project = self._make_project(
+            'TestProject', PROJECT_TYPE_PROJECT, self.category
+        )
+        self.owner_as = self._make_assignment(
+            self.project, self.user, self.role_owner
+        )
+
+    def test_post(self):
+        """Test POST for custom column retrieval"""
+        with self.login(self.user):
+            response = self.client.post(
+                reverse('projectroles:ajax_project_list_columns'),
+                json.dumps({'projects': [str(self.project.sodar_uuid)]}),
+                content_type='application/json',
+            )
+        self.assertEqual(response.status_code, 200)
+        expected = {
+            str(self.project.sodar_uuid): {
+                'filesfolders': {'files': {'html': '0'}, 'links': {'html': '0'}}
+            }
+        }
+        self.assertEqual(response.data, expected)
+
+    @override_settings(FILESFOLDERS_SHOW_LIST_COLUMNS=False)
+    def test_post_no_columns(self):
+        """Test POST with no custom colums"""
+        with self.login(self.user):
+            response = self.client.post(
+                reverse('projectroles:ajax_project_list_columns'),
+                json.dumps({'projects': [str(self.project.sodar_uuid)]}),
+                content_type='application/json',
+            )
+        self.assertEqual(response.status_code, 200)
+        expected = {str(self.project.sodar_uuid): {}}
+        self.assertEqual(response.data, expected)
+
+    def test_post_no_permission(self):
+        """Test POST with no user permission on a project"""
+        new_project = self._make_project(
+            'NewProject', PROJECT_TYPE_PROJECT, None
+        )
+        new_user = self.make_user('new_user')
+        self._make_assignment(new_project, new_user, self.role_owner)
+
+        with self.login(new_user):
+            response = self.client.post(
+                reverse('projectroles:ajax_project_list_columns'),
+                json.dumps(
+                    {
+                        'projects': [
+                            str(self.project.sodar_uuid),
+                            str(new_project.sodar_uuid),
+                        ]
+                    }
+                ),
+                content_type='application/json',
+            )
+        self.assertEqual(response.status_code, 200)
+        expected = {
+            str(new_project.sodar_uuid): {
+                'filesfolders': {'files': {'html': '0'}, 'links': {'html': '0'}}
+            }
+        }
+        self.assertEqual(response.data, expected)
 
 
 class TestProjectStarringAjaxView(
     ProjectMixin, RoleAssignmentMixin, ProjectUserTagMixin, TestViewsBase
 ):
-    """Tests for the project starring Ajax view"""
+    """Tests for ProjectStarringAjaxView"""
 
     def setUp(self):
         super().setUp()
@@ -29,11 +115,8 @@ class TestProjectStarringAjaxView(
 
     def test_star_project(self):
         """Test project starring"""
-
-        # Assert precondition
         self.assertEqual(ProjectUserTag.objects.all().count(), 0)
 
-        # Issue request
         with self.login(self.user):
             response = self.client.post(
                 reverse(
@@ -42,14 +125,11 @@ class TestProjectStarringAjaxView(
                 )
             )
 
-        # Assert ProjectUserTag state after creation
         self.assertEqual(ProjectUserTag.objects.all().count(), 1)
-
         tag = ProjectUserTag.objects.get(
             project=self.project, user=self.user, name=PROJECT_TAG_STARRED
         )
         self.assertIsNotNone(tag)
-
         expected = {
             'id': tag.pk,
             'project': self.project.pk,
@@ -57,20 +137,14 @@ class TestProjectStarringAjaxView(
             'name': PROJECT_TAG_STARRED,
             'sodar_uuid': tag.sodar_uuid,
         }
-
         self.assertEqual(model_to_dict(tag), expected)
-
-        # Assert redirect
         self.assertEqual(response.status_code, 200)
 
     def test_unstar_project(self):
         """Test project unstarring"""
         self._make_tag(self.project, self.user, name=PROJECT_TAG_STARRED)
-
-        # Assert precondition
         self.assertEqual(ProjectUserTag.objects.all().count(), 1)
 
-        # Issue request
         with self.login(self.user):
             response = self.client.post(
                 reverse(
@@ -79,8 +153,5 @@ class TestProjectStarringAjaxView(
                 )
             )
 
-        # Assert ProjectUserTag state after creation
         self.assertEqual(ProjectUserTag.objects.all().count(), 0)
-
-        # Assert status code
         self.assertEqual(response.status_code, 200)