+++ projectroles/urls.py
@@ -143,6 +143,11 @@ urls_ui = [
 # Ajax API views
 urls_ajax = [
     url(
+        regex=r'^ajax/list/columns',
+        view=views_ajax.ProjectListColumnAjaxView.as_view(),
+        name='ajax_project_list_columns',
+    ),
+    url(
         regex=r'^ajax/star/(?P<project>[0-9a-f-]+)',
         view=views_ajax.ProjectStarringAjaxView.as_view(),
         name='ajax_star',